+++ CHANGELOG.md
@@ -15,6 +15,7 @@ This CHANGELOG follows the format located [here](https://github.com/sensu-plugin
 - `filters` resource now supports `runtime_assets` being passed (@majormoses)
 - Added support for `postgres_config` resource (@cwjohnston) - #70
 - Added support for `role`, `role_binding`, `cluster_role` and `cluster_role_binding` resources (@cwjohnston) - #71
+- Added init action to sensu-backend to fix functionality added in 5.16.0 - #77
 
 ### Breaking changes